pipeline_faiss/requirements.txt
PyQt6==6.7.1
torch
transformers
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy
pypdf==4.3.1
chromadb==1.0.20
pymupdf==1.24.9
pdfminer.six==20240706
